+++ src/js/language/hu_HU.js
@@ -229,6 +229,7 @@
             country: 'Kérlek, hogy érvényes %s telefonszámot adj meg',
             countries: {
                 GB: 'Egyesült Királyság beli',
+                FR: 'francia',
                 US: 'Egyesült Államok beli'
             }
         },