+++ CHANGELOG.md
@@ -6,6 +6,7 @@ __New features__
 
 * [#218](https://github.com/nghuuphuoc/bootstrapvalidator/issues/218), [#531](https://github.com/nghuuphuoc/bootstrapvalidator/pull/531): Add meid validator, thanks to [@troymccabe](https://github.com/troymccabe)
 * [#267](https://github.com/nghuuphuoc/bootstrapvalidator/issues/267), [#532](https://github.com/nghuuphuoc/bootstrapvalidator/pull/532): Add imo validator, thanks to [@troymccabe](https://github.com/troymccabe)
+* [#510](https://github.com/nghuuphuoc/bootstrapvalidator/pull/510): The [phone](http://bootstrapvalidator.com/validators/phone/) validator now supports French phone number, thanks to [@dlucazeau](https://github.com/dlucazeau)
 
 __Improvements__